+++ helpers/test-client/src/sc/ScoreValue.kt
@@ -1,0 +1,6 @@
+package sc
+
+import sc.shared.ScoreFragment
+import java.math.BigDecimal
+
+data class ScoreValue(val fragment: ScoreFragment, var value: BigDecimal)