+++ sdk/src/main/server-api/sc/networking/InvalidScoreDefinitionException.java
@@ -1,9 +1,0 @@
-package sc.networking;
-
-public class InvalidScoreDefinitionException extends Exception {
-
-  public InvalidScoreDefinitionException(String msg) {
-    super(msg);
-  }
-
-}